% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNetwork.R
\name{createVisNetwork}
\alias{createVisNetwork}
\title{visNetwork}
\usage{
createVisNetwork(
  nodes,
  edges,
  background = "lightblue",
  border = "darkblue",
  highlight = "yellow"
)
}
\arguments{
\item{nodes}{Nodes.}

\item{edges}{Edges.}

\item{background}{Background color.}

\item{border}{Border color.}

\item{highlight}{Highlight color.}
}
\value{
Network map visualization
}
\description{
visNetwork
}
\examples{
nodes <- getNodes(head(mtcars), c("cyl", "gear", "vs"), group = TRUE)
res <- fixNodeBias(head(mtcars))
edges <- getEdges(getAssociation(res), getNodes(res, group = TRUE))
createVisNetwork(nodes, edges)
}
\author{
Jayachandra N
}
