% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_association.R
\name{getAssociation}
\alias{getAssociation}
\title{getAssociation}
\usage{
getAssociation(crude_data)
}
\arguments{
\item{crude_data}{data frame to get associations between the values of different columns}
}
\value{
data frame of edges indicating from and to nodes
}
\description{
getAssociation
}
\examples{
\dontrun{
getAssociation(head(mtcars))
}
}
\author{
Jayachandra N
}
