% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdges.R
\name{getEdges}
\alias{getEdges}
\title{getEdges}
\usage{
getEdges(polished_data, nodes)
}
\arguments{
\item{polished_data}{data frame, output of getAssociation function}

\item{nodes}{data frame, output of getNodes function}
}
\value{
data frame of edges indicationg from and to node ids
}
\description{
Generate edges or lines data frame which defines the link between nodes.
}
\examples{
res <- fixNodeBias(head(mtcars))
edges <- getEdges(getAssociation(res), getNodes(res, group = TRUE))
}
\author{
Jayachandra N
}
