% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixNodeBias.R
\name{fixNodeBias}
\alias{fixNodeBias}
\title{fixNodeBias}
\usage{
fixNodeBias(my_data)
}
\arguments{
\item{my_data}{data frame}
}
\value{
data frame
}
\description{
Fix the possible bias that occurs while generating the nodes.
}
\examples{
res <- fixNodeBias(head(mtcars))
}
\author{
Jayachandra N
}
