% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShapes.R
\name{getShapes}
\alias{getShapes}
\title{getShapes}
\usage{
getShapes(nodes, shapes = NULL)
}
\arguments{
\item{nodes}{nodes data, output of getNodes function}

\item{shapes}{character vector indicating manual shapes to choose for nodes}
}
\value{
a vector of shapes which is ready to cbind with nodes data
}
\description{
Generate shapes for the given nodes, shapes can be defined using the
shapes otherwise defaults to null.
}
\examples{
nodes <- getNodes(head(mtcars), c("cyl", "gear", "vs"), group =TRUE)
nodes$shape <- getShapes(nodes)
}
\author{
Jayachandra N
}
