% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitNet.R
\name{knitNet}
\alias{knitNet}
\title{knitNet}
\usage{
knitNet()
}
\value{
shiny ui page
}
\description{
Run inbuilt shiny app after installing the package.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
knitNet()
\dontshow{\}) # examplesIf}
}
\author{
Jayachandra N
}
