% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMyFile.R
\name{readMyFile}
\alias{readMyFile}
\title{readMyFile}
\usage{
readMyFile(this_file)
}
\arguments{
\item{this_file}{File path}
}
\value{
data.frame, content of the file.
}
\description{
readMyFile
}
\examples{
write.csv(mtcars, 'test_data.csv')
readMyFile("test_data.csv")
}
\author{
Jayachandra N
}
